requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
soundfile
numpy==2.0.1
torchmetrics
torchmetrics[audio]
torch_audiomentations
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core
gdown
opencv-python

black
isort
pre-commit
wget
thop
gdown
opencv-python
yadisk
